refactor: move doctor command from root to spec namespace
Doctor diagnoses .vaultspec/ framework health, which is semantically a
spec operation. Moving it under `spec doctor` aligns with the vault/spec
namespace separation established in #36.

- Move cmd_doctor and helpers (_signal_status, _provider_status,
  _doctor_exit_code) from cli/root.py to cli/spec_cmd.py
- Remove unused debug parameter (handled by root callback)
- Update spec-check pre-commit hook entry to `spec doctor`
- Update all test invocations to use new path

Closes #46
