+++ src/vaultspec_core/tests/cli/test_doctor.py
@@ -1,4 +1,4 @@
-"""Tests for the ``vaultspec-core doctor`` CLI command."""
+"""Tests for the ``vaultspec-core spec doctor`` CLI command."""
 
 from __future__ import annotations
 
@@ -21,25 +21,25 @@ class TestDoctorCommand:
     def test_installed_workspace_does_not_error(self, tmp_path: Path) -> None:
         factory = WorkspaceFactory(tmp_path)
         factory.install()
-        result = factory.run("doctor")
+        result = factory.run("spec", "doctor")
         assert result.exit_code in (0, 1)
 
     def test_output_contains_framework(self, tmp_path: Path) -> None:
         factory = WorkspaceFactory(tmp_path)
         factory.install()
-        result = factory.run("doctor")
+        result = factory.run("spec", "doctor")
         assert "framework" in result.output.lower()
 
     def test_corrupted_manifest_exit_two(self, tmp_path: Path) -> None:
         factory = WorkspaceFactory(tmp_path)
         factory.install().corrupt_manifest()
-        result = factory.run("doctor")
+        result = factory.run("spec", "doctor")
         assert result.exit_code == 2
 
     def test_json_output_valid(self, tmp_path: Path) -> None:
         factory = WorkspaceFactory(tmp_path)
         factory.install()
-        result = factory.run("doctor", "--json")
+        result = factory.run("spec", "doctor", "--json")
         data = json.loads(result.output)
         assert "framework" in data
         assert "providers" in data
@@ -50,7 +50,7 @@ class TestDoctorCommand:
     def test_json_exit_code_reflects_corrupted_state(self, tmp_path: Path) -> None:
         factory = WorkspaceFactory(tmp_path)
         factory.install().corrupt_manifest()
-        result = factory.run("doctor", "--json")
+        result = factory.run("spec", "doctor", "--json")
         assert result.exit_code == 2
         # The output may contain log warnings before the JSON payload.
         json_start = result.output.index("{")
@@ -59,42 +59,42 @@ class TestDoctorCommand:
 
     def test_missing_framework_exit_two(self, tmp_path: Path) -> None:
         factory = WorkspaceFactory(tmp_path)
-        result = factory.run("doctor")
+        result = factory.run("spec", "doctor")
         assert result.exit_code == 2
 
     def test_output_contains_provider_names(self, tmp_path: Path) -> None:
         factory = WorkspaceFactory(tmp_path)
         factory.install()
-        result = factory.run("doctor")
+        result = factory.run("spec", "doctor")
         assert "claude" in result.output.lower()
 
     def test_output_contains_builtins_row(self, tmp_path: Path) -> None:
         factory = WorkspaceFactory(tmp_path)
         factory.install()
-        result = factory.run("doctor")
+        result = factory.run("spec", "doctor")
         assert "builtins" in result.output.lower()
 
     def test_output_contains_gitignore_row(self, tmp_path: Path) -> None:
         factory = WorkspaceFactory(tmp_path)
         factory.install()
-        result = factory.run("doctor")
+        result = factory.run("spec", "doctor")
         assert "gitignore" in result.output.lower()
 
     def test_output_contains_gitattributes_row(self, tmp_path: Path) -> None:
         factory = WorkspaceFactory(tmp_path)
         factory.install()
-        result = factory.run("doctor")
+        result = factory.run("spec", "doctor")
         assert "gitattributes" in result.output.lower()
 
     def test_deleted_vaultspec_dir_exit_two(self, tmp_path: Path) -> None:
         factory = WorkspaceFactory(tmp_path)
         factory.install().delete_vaultspec_dir()
-        result = factory.run("doctor")
+        result = factory.run("spec", "doctor")
         assert result.exit_code == 2
 
     def test_json_healthy_framework_present(self, tmp_path: Path) -> None:
         factory = WorkspaceFactory(tmp_path)
         factory.install()
-        result = factory.run("doctor", "--json")
+        result = factory.run("spec", "doctor", "--json")
         data = json.loads(result.output)
         assert data["framework"] == "present"
